software/camcom/requirements.txt
cmake
opencv-python
numpy
picamera2
apriltag
cmapy
smbus
crcmod
matplotlib
pyserial
shapely